# services/block-i-signals/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Block I Activity Ingestion and Signal Service
  version: 1.0.0
  description: >
    Ingests activity events, enforces privacy thresholds and retention,
    and serves per-user and aggregate document signals.
servers:
  - url: http://localhost:8089
paths:
  /activity/ingest:
    post:
      summary: Ingest activity events
      operationId: ingestActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Accepted (or partial)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '401':
          description: Unauthorized
        '403':
          description: Insufficient scope
  /signals/user/{user_id}:
    get:
      summary: Per-user signal vector
      operationId: getUserSignals
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignalResponse'
  /signals/document/{document_id}:
    get:
      summary: Aggregate document popularity (privacy gated)
      operationId: getDocumentSignals
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document signals or privacy_protected payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSignalResponse'
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ActivityEvent:
      type: object
      required: [event_id, actor_principal_id, object_id, event_type, source_system, event_time]
      properties:
        event_id: { type: string }
        tenant_id: { type: string, description: Ignored; taken from JWT }
        actor_principal_id: { type: string }
        object_id: { type: string }
        event_type:
          type: string
          enum: [view, edit, authored, commented_on, referenced, worked_on]
        source_system: { type: string }
        event_time: { type: string, format: date-time }
        session_id: { type: string }
        context_json: { type: object }
        privacy_level:
          type: string
          enum: [public, restricted, confidential]
        ttl_seconds: { type: integer }
    IngestRequest:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/ActivityEvent'
    IngestResponse:
      type: object
      properties:
        status: { type: string }
        ingested_count: { type: integer }
        already_processed_count: { type: integer }
        failed_events:
          type: array
          items:
            type: object
            properties:
              event_id: { type: string }
              reason: { type: string }
    UserSignalResponse:
      type: object
      properties:
        user_id: { type: string }
        tenant_id: { type: string }
        signals: { type: object }
        updated_at: { type: string, format: date-time }
        freshness_s: { type: number, nullable: true }
    DocumentSignalResponse:
      type: object
      properties:
        document_id: { type: string }
        tenant_id: { type: string }
        privacy_protected: { type: boolean }
        popularity_score: { type: number, nullable: true }
        total_views: { type: integer, nullable: true }
        distinct_viewers: { type: integer, nullable: true }
        last_viewed: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
        window_days: { type: integer, nullable: true }